268 Refactor main-map; fix routing bug with redirect
