+++ app/templates/application.hbs
@@ -73,3 +73,6 @@
     {{outlet}}
   </div>
 </div>
+
+
+{{outlet}}
