+++ tests/acceptance/navigating-to-qpd-url-breaks-test.js
@@ -34,6 +34,7 @@ module('Acceptance | navigating to qpd url breaks', function(hooks) {
 
     // await percySnapshot('qp test: visit index, change some things');
 
+    // the refresh here resets the context, requiring re-mocking
     await refresh();
 
     // await percySnapshot('after a refresh, things are applied from previous QPs');
